src/main/python/requirements.txt
Pillow>=9.5.0
GDAL>=3.6.0
numpy>=1.24.0 